+++ docs/release-notes.md
@@ -6,17 +6,29 @@ hide:
 
 # Release Notes
 
-## Unreleased
+## 0.20.0
 
 ### Added
 
 - Add DurationField.
 - Allow passing `max_digits` to FloatField.
+- Add `local_or` function to QuerySets.
+
+### Changed
+
+- Only the main table of a queryset is queryable via `model_class.columns.foo == foo`. Select related models have now an unique name for their path.
+  The name can be retrieved via `tables_and_models` or using `f"{hash_tablekey(...)}_{column}"`.
+- Breaking: Alter tables_and_models to use the prefix as key with '' for the maintable and model.
+- Breaking: Functions passed to filter functions reveive now a second positional parameter `tables_and_models`.
+- `build_where_clause` conditionally uses a subquery.
+- Rename QueryType to QuerySetType. The old name stays as an alias.
 
 ### Fixed
 
 - Triggering load on non-existent field when reflecting.
 - InspectDB mapping was incorrect.
+- Fix query edge cases.
+- Fix using related queries with update/delete.
 
 
 ## 0.19.1